SilverStripe 4 support
