+++ _config.php
@@ -1,6 +1,2 @@
 <?php
 
-if (!defined('TEXTTARGETLENGTH_DIR')) {
-    //Assumes this module is installed in web root
-    define('TEXTTARGETLENGTH_DIR', basename(dirname(__FILE__)));
-}
